feat: Introduce Knowledge schemas
These describe the schema for the knowledge graph to be extracted.
Currently, they allow defining the types of nodes and relationships
that can be extracted. We could extend this in the future to allow
properties of nodes/edges to be extracted and merged, etc.

This uses a prompt loaded from `extraction.md` for easy iteration.