feat: Add a properties map to nodes
The current implementation is a bit hacky -- the node hash and identity
doesn't consider the property dict, because in some places we create
a dictionary/set keyed by node. What we probably need is a `NodeId` or
`NodeKey` that we can use when we mean "the identity of the node" not
the "full value of the node". This should work for now.